+++ naplib/io/__init__.py
@@ -1,11 +1,12 @@
-from .fileio import load, save, import_data, read_bids, export_data
+from .fileio import load, save, import_data, export_data
+from .load_bids import load_bids
 from .load_sample import load_speech_task_data
 from .read_htk import read_htk
-from .read_cnd import read_cnd
+from .load_cnd import load_cnd
 from .load_tdt import load_tdt
 from .load_nwb import load_nwb
 from .load_edf import load_edf
 from .load_wav_dir import load_wav_dir
 
-__all__ = ['load','save','import_data','export_data','read_bids', 'read_cnd',
+__all__ = ['load','save','import_data','export_data','load_bids', 'load_cnd',
            'load_speech_task_data','read_htk','load_tdt','load_nwb','load_edf','load_wav_dir']
